Fix Build and HTTP Parser Call (#46)
The HTTP gem changed its underlying http parser library to a different
gem, and that broke the streaming response code that is in this gem,
causing an error when trying to load it. This fix uses the HTTP gem's
streaming API, not `http-parser` itself, so it should be a little more
immune from these breaking changes. Also add a missing expectation in
the connection tests.